+++ Infrastructure/Conventions/TestControllersConvention.cs
@@ -1,10 +1,18 @@
 ﻿using System;
+using Microsoft.AspNetCore.Mvc.ApplicationModels;
+
 namespace WebStore_MVC.Infrastructure.Conventions
 {
-    public class TestControllersConvention
+    public class TestControllersConvention:IControllerModelConvention
     {
         public TestControllersConvention()
         {
+        }
+
+        public void Apply(ControllerModel controller)
+        {
+            //controller.Actions.Add(new ActionModel());
+            //throw new NotImplementedException();
         }
     }
 }
