+++ Views/Employees/Details.cshtml
@@ -39,6 +39,6 @@
     </div>
     <div>
         @Html.ActionLink("Edit", "Edit", new { /* id = Model.PrimaryKey */ }) |
-        <a asp-action="Employees">Back to List</a>
+        <a asp-action="Index">Back to List</a>
     </div>
 </div>
