+++ Views/Shared/Partial/_MainMenu.cshtml
@@ -20,6 +20,6 @@
         </li>
         <li><a asp-controller="Home" asp-action="NotFound404">404</a></li>
         <li><a asp-controller="Home" asp-action="ContactUs">Contact</a></li>
-        <li><a asp-controller="Home" asp-action="Employees">Our Team</a></li>
+        <li><a asp-controller="Employees" asp-action="Index">Our Team</a></li>
     </ul>
 </div>
