+++ WebStore_MVC.csproj
@@ -32,5 +32,6 @@
     <Folder Include="Infrastructure\" />
     <Folder Include="Infrastructure\MiddleWare\" />
     <Folder Include="Infrastructure\Conventions\" />
+    <Folder Include="Views\Employees\" />
   </ItemGroup>
 </Project>
